MultiModel_AiProject/requirements.txt
chainlit==1.1.401
python-dotenv==1.0.1
langchain==0.2.11
langchain-groq==0.1.5
groq==0.9.0
httpx>=0.23.0
